Lint + create Utils singleton
