+++ src/initialize-singletons.js
@@ -20,6 +20,7 @@ import { PropertiesImplementation } from "./methods/properties.js";
 import { ToImplementation } from "./methods/to.js";
 import { WidenImplementation } from "./methods/widen.js";
 import { concretize } from "./utils/ConcreteModelConverter.js";
+import * as utils from "./utils.js";
 
 export default function() {
   Singletons.setCreate(new CreateImplementation());
@@ -32,4 +33,5 @@ export default function() {
   Singletons.setTo((new ToImplementation(): any));
   Singletons.setWiden((new WidenImplementation(): any));
   Singletons.setConcretize(concretize);
+  Singletons.setUtils(utils);
 }
