+++ src/intrinsics/prepack/utils.js
@@ -25,6 +25,7 @@ import { describeLocation } from "../ecma262/Error.js";
 import { To } from "../../singletons.js";
 import AbstractObjectValue from "../../values/AbstractObjectValue";
 import { CompilerDiagnostic, FatalError } from "../../errors.js";
+import { Utils } from "../../singletons";
 
 const throwTemplateSrc = "(function(){throw new global.Error('abstract value defined at ' + A);})()";
 const throwTemplate = buildExpressionTemplate(throwTemplateSrc);
@@ -36,14 +37,14 @@ export function parseTypeNameOrTemplate(
   if (typeNameOrTemplate === undefined || typeNameOrTemplate instanceof UndefinedValue) {
     return { type: Value, template: undefined };
   } else if (typeof typeNameOrTemplate === "string") {
-    let type = Value.getTypeFromName(typeNameOrTemplate);
+    let type = Utils.getTypeFromName(typeNameOrTemplate);
     if (type === undefined) {
       throw realm.createErrorThrowCompletion(realm.intrinsics.TypeError, "unknown typeNameOrTemplate");
     }
     return { type, template: undefined };
   } else if (typeNameOrTemplate instanceof StringValue) {
     let typeNameString = To.ToStringPartial(realm, typeNameOrTemplate);
-    let type = Value.getTypeFromName(typeNameString);
+    let type = Utils.getTypeFromName(typeNameString);
     if (type === undefined) {
       throw realm.createErrorThrowCompletion(realm.intrinsics.TypeError, "unknown typeNameOrTemplate");
     }
