+++ src/types.js
@@ -1007,3 +1007,8 @@ export type ToType = {
 };
 
 export type ConcretizeType = (realm: Realm, val: Value) => ConcreteValue;
+
+export type UtilsType = {|
+  typeToString: (typeof Value) => void | string,
+  getTypeFromName: string => void | typeof Value,
+|};
