+++ src/utils.js
@@ -1,0 +1,85 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+/* @flow */
+
+import {
+  AbstractObjectValue,
+  ArrayValue,
+  BooleanValue,
+  ConcreteValue,
+  EmptyValue,
+  FunctionValue,
+  NullValue,
+  NumberValue,
+  IntegralValue,
+  ObjectValue,
+  PrimitiveValue,
+  StringValue,
+  SymbolValue,
+  UndefinedValue,
+  Value
+} from "./values/index.js";
+import invariant from './invariant';
+
+export function typeToString(type: typeof Value): void | string {
+  function isInstance(proto, Constructor): boolean {
+    return proto instanceof Constructor || proto === Constructor.prototype;
+  }
+  let proto = type.prototype;
+  if (isInstance(proto, UndefinedValue)) {
+    return "undefined";
+  } else if (isInstance(proto, NullValue)) {
+    return "object";
+  } else if (isInstance(proto, StringValue)) {
+    return "string";
+  } else if (isInstance(proto, BooleanValue)) {
+    return "boolean";
+  } else if (isInstance(proto, NumberValue)) {
+    return "number";
+  } else if (isInstance(proto, SymbolValue)) {
+    return "symbol";
+  } else if (isInstance(proto, ObjectValue)) {
+    if (Value.isTypeCompatibleWith(type, FunctionValue)) {
+      return "function";
+    }
+    return "object";
+  } else {
+    return undefined;
+  }
+}
+
+export function getTypeFromName(typeName: string): void | typeof Value {
+  switch (typeName) {
+    case "empty":
+      return EmptyValue;
+    case "void":
+      return UndefinedValue;
+    case "null":
+      return NullValue;
+    case "boolean":
+      return BooleanValue;
+    case "string":
+      return StringValue;
+    case "symbol":
+      return SymbolValue;
+    case "number":
+      return NumberValue;
+    case "object":
+      return ObjectValue;
+    case "array":
+      return ArrayValue;
+    case "function":
+      return FunctionValue;
+    case "integral":
+      return IntegralValue;
+    default:
+      return undefined;
+  }
+}
